move peak picker to centroiding in filtering
