+++ src/openms/include/OpenMS/FILTERING/CENTROIDING/PeakPickerIterative.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include <OpenMS/TRANSFORMATIONS/CENTROIDING/PeakPickerHiRes.h>
+#include <OpenMS/FILTERING/CENTROIDING/PeakPickerHiRes.h>
 #include <OpenMS/FILTERING/NOISEESTIMATION/SignalToNoiseEstimatorMedian.h>
 #include <OpenMS/CONCEPT/ProgressLogger.h>
 #include <OpenMS/KERNEL/SpectrumHelper.h>
