+++ src/openms/include/OpenMS/FILTERING/CENTROIDING/sources.cmake
@@ -1,5 +1,5 @@
 ### the directory name
-set(directory include/OpenMS/TRANSFORMATIONS/CENTROIDING)
+set(directory include/OpenMS/FILTERING/CENTROIDING)
 
 ### list all header files of the directory here
 set(sources_list_h
@@ -14,7 +14,7 @@ foreach(i ${sources_list_h})
 endforeach(i)
 
 ### source group definition
-source_group("Header Files\\OpenMS\\TRANSFORMATIONS\\CENTROIDING" FILES ${sources_h})
+source_group("Header Files\\OpenMS\\FILTERING\\CENTROIDING" FILES ${sources_h})
 
 set(OpenMS_sources_h ${OpenMS_sources_h} ${sources_h})
