+++ src/openms/includes.cmake
@@ -42,7 +42,7 @@ include(source/FILTERING/SMOOTHING/sources.cmake)
 include(source/FILTERING/NOISEESTIMATION/sources.cmake)
 include(source/FILTERING/ID/sources.cmake)
 include(source/TRANSFORMATIONS/FEATUREFINDER/sources.cmake)
-include(source/TRANSFORMATIONS/CENTROIDING/sources.cmake)
+include(source/FILTERING/CENTROIDING/sources.cmake)
 include(source/COMPARISON/SPECTRA/sources.cmake)
 include(source/QC/sources.cmake)
 if(NOT DISABLE_OPENSWATH)
@@ -98,7 +98,7 @@ include(include/OpenMS/FILTERING/NOISEESTIMATION/sources.cmake)
 include(include/OpenMS/FILTERING/SMOOTHING/sources.cmake)
 include(include/OpenMS/FILTERING/TRANSFORMERS/sources.cmake)
 include(include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/sources.cmake)
-include(include/OpenMS/TRANSFORMATIONS/CENTROIDING/sources.cmake)
+include(include/OpenMS/FILTERING/CENTROIDING/sources.cmake)
 include(include/OpenMS/QC/sources.cmake)
 if(NOT DISABLE_OPENSWATH)
   include(include/OpenMS/ANALYSIS/OPENSWATH/sources.cmake)
