+++ src/openms/source/ANALYSIS/OPENSWATH/SwathQC.cpp
@@ -12,7 +12,7 @@
 #include <OpenMS/CONCEPT/Exception.h>
 #include <OpenMS/FILTERING/DATAREDUCTION/Deisotoper.h>
 #include <OpenMS/METADATA/ExperimentalSettings.h>
-#include <OpenMS/TRANSFORMATIONS/CENTROIDING/PeakPickerHiRes.h>
+#include <OpenMS/FILTERING/CENTROIDING/PeakPickerHiRes.h>
 
 #include <nlohmann/json.hpp>
