+++ src/openms/source/ANALYSIS/XLMS/OpenPepXLLFAlgorithm.cpp
@@ -22,7 +22,7 @@
 #include <OpenMS/CHEMISTRY/TheoreticalSpectrumGeneratorXLMS.h>
 #include <OpenMS/KERNEL/SpectrumHelper.h>
 #include <OpenMS/MATH/MISC/MathFunctions.h>
-#include <OpenMS/TRANSFORMATIONS/CENTROIDING/PeakPickerHiRes.h>
+#include <OpenMS/FILTERING/CENTROIDING/PeakPickerHiRes.h>
 
 #include <iostream>
