+++ src/tests/class_tests/openms/source/PeakPickerHiRes_test.cpp
@@ -11,7 +11,7 @@
 #include <OpenMS/FORMAT/MzMLFile.h>
 
 ///////////////////////////
-#include <OpenMS/TRANSFORMATIONS/CENTROIDING/PeakPickerHiRes.h>
+#include <OpenMS/FILTERING/CENTROIDING/PeakPickerHiRes.h>
 ///////////////////////////
 
 using namespace OpenMS;
